Fix linter: extend used outside of class or module
