+++ main.rb
@@ -1,11 +1,22 @@
 require_relative 'school_library/app'
-require_relative 'school_library/modules/display_menu'
 def main
   App.run
 end
 
+def display_menu
+  puts "\nPlease choose an option by enter a number:"
+  puts '1. List all books'
+  puts '2. List all people'
+  puts '3. Create a person'
+  puts '4. Create a book'
+  puts '5. Create a rental'
+  puts '6. List all rentals for a given person (by ID)'
+  puts '7. Exit'
+  print 'Enter your choice: '
+  gets.chomp.to_i
+end
+
 def handle_choice(choice)
-  extend DisplayMenu
   case choice
   when 1
     list_books
